gh-2675 split the ToCsv into individual operations depending on format required, addded correspodning handler and test
